# zhiyin-src/template/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponse_DeveloperProjectView_": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeveloperProjectView"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "ok",
            "title": "Message",
            "type": "string"
          },
          "trace_id": {
            "description": "链路追踪 id，由 BFF 生成并回写 X-Trace-Id 响应头；日志排查用",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "title": "ApiResponse[DeveloperProjectView]",
        "type": "object"
      },
      "ApiResponse_DeveloperVersionView_": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeveloperVersionView"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "ok",
            "title": "Message",
            "type": "string"
          },
          "trace_id": {
            "description": "链路追踪 id，由 BFF 生成并回写 X-Trace-Id 响应头；日志排查用",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "title": "ApiResponse[DeveloperVersionView]",
        "type": "object"
      },
...
      "ApiResponse_dict_": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "default": "ok",
            "title": "Message",
            "type": "string"
          },
          "trace_id": {
            "description": "链路追踪 id，由 BFF 生成并回写 X-Trace-Id 响应头；日志排查用",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "title": "ApiResponse[dict]",
        "type": "object"
      },
...
      "ApiResponse_list_DeveloperProjectView__": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DeveloperProjectView"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "default": "ok",
            "title": "Message",
            "type": "string"
          },
          "trace_id": {
            "description": "链路追踪 id，由 BFF 生成并回写 X-Trace-Id 响应头；日志排查用",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "title": "ApiResponse[list[DeveloperProjectView]]",
        "type": "object"
      },
      "ApiResponse_list_DeveloperVersionView__": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "default": 0
          },
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DeveloperVersionView"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "message": {
            "default": "ok",
            "title": "Message",
            "type": "string"
          },
          "trace_id": {
            "description": "链路追踪 id，由 BFF 生成并回写 X-Trace-Id 响应头；日志排查用",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "title": "ApiResponse[list[DeveloperVersionView]]",
        "type": "object"
      },
...
      "DeveloperProjectCreate": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "default": "",
            "maxLength": 1000,
            "title": "Description",
            "type": "string"
          },
          "id": {
            "pattern": "^[a-z][a-z0-9_]{1,47}$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "DeveloperProjectCreate",
        "type": "object"
      },
      "DeveloperProjectUpdate": {
        "additionalProperties": false,
        "properties": {
          "auto_deploy": {
            "default": false,
            "title": "Auto Deploy",
            "type": "boolean"
          },
          "description": {
            "default": "",
            "maxLength": 1000,
            "title": "Description",
            "type": "string"
          },
          "members": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "title": "Members",
            "type": "array"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "revision": {
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "trusted": {
            "default": false,
            "title": "Trusted",
            "type": "boolean"
          }
        },
        "required": [
          "revision",
          "name"
        ],
        "title": "DeveloperProjectUpdate",
        "type": "object"
      },
...
      "DeveloperUpload": {
        "additionalProperties": false,
        "properties": {
          "base_commit": {
            "pattern": "^[a-f0-9]{40}$",
            "title": "Base Commit",
            "type": "string"
          },
          "base_version_id": {
            "anyOf": [
              {
                "pattern": "^[a-f0-9]{32}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Version Id"
          },
          "channel": {
            "default": "main",
            "pattern": "^[a-z][a-z0-9_-]{0,39}$",
            "title": "Channel",
            "type": "string"
          },
          "package_base64": {
            "maxLength": 2800000,
            "minLength": 1,
            "title": "Package Base64",
            "type": "string"
          },
          "request_id": {
            "maxLength": 80,
            "minLength": 8,
            "pattern": "^[a-zA-Z0-9_-]+$",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "package_base64",
          "base_commit",
          "request_id"
        ],
        "title": "DeveloperUpload",
        "type": "object"
      },
...
      "ErrorCode": {
        "description": "统一错误码。前端据此提示，不解析后端文案。",
        "enum": [
          0,
          1001,
          1002,
          1003,
          1004,
          1005,
          1006,
          1007,
          1999
        ],
        "title": "ErrorCode",
        "type": "integer"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "职引 API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/developer/executor/{job_id}/authorization": {
      "get": {
        "operationId": "release_authorization_api_v1_developer_executor__job_id__authorization_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release Authorization",
        "tags": [
          "developer-projects"
        ]
      }
    },
...
    "/api/v1/developer/projects": {
      "get": {
        "operationId": "projects_api_v1_developer_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_DeveloperProjectView__"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Projects",
        "tags": [
          "developer-projects"
        ]
      },
      "post": {
        "operationId": "create_project_api_v1_developer_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeveloperProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DeveloperProjectView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "developer-projects"
        ]
      }
    },
    "/api/v1/developer/projects/{project_id}": {
      "put": {
        "operationId": "update_project_api_v1_developer_projects__project_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeveloperProjectUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DeveloperProjectView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Project",
        "tags": [
          "developer-projects"
        ]
      }
    },
    "/api/v1/developer/projects/{project_id}/versions": {
      "get": {
        "operationId": "versions_api_v1_developer_projects__project_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_DeveloperVersionView__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Versions",
        "tags": [
          "developer-projects"
        ]
      },
      "post": {
        "operationId": "upload_version_api_v1_developer_projects__project_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeveloperUpload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DeveloperVersionView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Version",
        "tags": [
          "developer-projects"
        ]
      }
    },
...
    "/api/v1/developer/source-worker/heartbeat": {
      "post": {
        "operationId": "heartbeat_api_v1_developer_source_worker_heartbeat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeveloperWorkerHeartbeat"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Heartbeat",
        "tags": [
          "developer-projects"
        ]
      }
    },
...
    "/api/v1/developer/source-worker/versions/{version_id}": {
      "get": {
        "operationId": "worker_version_api_v1_developer_source_worker_versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DeveloperVersionView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Worker Version",
        "tags": [
          "developer-projects"
        ]
      },
      "post": {
        "operationId": "worker_update_api_v1_developer_source_worker_versions__version_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeveloperWorkerUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DeveloperVersionView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Worker Update",
        "tags": [
          "developer-projects"
        ]
      }
    },
    "/api/v1/developer/source-worker/versions/{version_id}/package": {
      "get": {
        "operationId": "worker_package_api_v1_developer_source_worker_versions__version_id__package_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Worker Package",
        "tags": [
          "developer-projects"
        ]
      }
    },
    "/api/v1/developer/source-worker/versions/{version_id}/release": {
      "post": {
        "operationId": "worker_release_api_v1_developer_source_worker_versions__version_id__release_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ReleaseJob_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Worker Release",
        "tags": [
          "developer-projects"
        ]
      }
    },
    "/api/v1/developer/source-worker/versions/{version_id}/retry": {
      "post": {
        "operationId": "worker_retry_api_v1_developer_source_worker_versions__version_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DeveloperVersionView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Worker Retry",
        "tags": [
          "developer-projects"
        ]
      }
    },
    "/api/v1/developer/templates": {
      "get": {
        "operationId": "template_api_v1_developer_templates_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "hybrid",
              "enum": [
                "application",
                "tool",
                "hybrid"
              ],
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "module_id",
            "required": true,
            "schema": {
              "pattern": "^[a-z][a-z0-9_]{1,47}$",
              "title": "Module Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 1,
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "owner",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 1,
              "title": "Owner",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Template",
        "tags": [
          "developer-projects"
        ]
      }
    },
    "/api/v1/developer/versions/{version_id}": {
      "get": {
        "operationId": "version_api_v1_developer_versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DeveloperVersionView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Version",
        "tags": [
          "developer-projects"
        ]
      }
    },
    "/api/v1/developer/versions/{version_id}/package": {
      "get": {
        "operationId": "package_api_v1_developer_versions__version_id__package_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Package",
        "tags": [
          "developer-projects"
        ]
      }
    },
    "/api/v1/developer/versions/{version_id}/release": {
      "post": {
        "operationId": "release_api_v1_developer_versions__version_id__release_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ReleaseJob_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release",
        "tags": [
          "developer-projects"
        ]
      }
    },
    "/api/v1/developer/versions/{version_id}/retry": {
      "post": {
        "operationId": "retry_api_v1_developer_versions__version_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DeveloperVersionView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry",
        "tags": [
          "developer-projects"
        ]
      }
    },
...
  }
}